requirements.txt
asgiref==3.2.7
backcall==0.2.0
certifi==2020.6.20
decorator==4.4.2
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
jupyter-client==6.1.3
jupyter-core==4.6.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.4
